Update user entity and create new ones
Initially I created a 'TempUser' entity for users who were invited to join a trip, so that they are not forced to register, just to participate in the trip they were invited in. However since they share similarities with a registered user, I just made the password field optional.

Additionally, I also had 'Restaurant', 'Hotel' and 'Attraction' entities, which turned to the 'Experience' entity, since they share a lot of similar fields. This was done because it would have made interacting with the 'Reaction' entity a bit tricky, most likely resulting in the creation of a separate Reaction entity for each experience or adding unnecessary optional values to the existing Reaction entity.